code for temperature observation
Add basic pseduo-code to show the desired process.  Also adds basic code
to utilize the temperature sensors used in this project.  Circuit
diagram to be added now that we have a circuit.

Proof of concept for temp sensor

Temperature sensor runs, seems to estimate 6/7 degrees low.  Not bad for
a cheap sensor.  Will require package tuning on complete installs or
better sensors.